% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_stdev_contrib.R
\name{asset_stdev_contrib}
\alias{asset_stdev_contrib}
\title{Asset contribution to portfolio standard deviation}
\usage{
asset_stdev_contrib(w, s, C, marginal = FALSE)
}
\arguments{
\item{w}{vector}

\item{s}{vector}

\item{C}{matrix}

\item{marginal}{bool}
}
\value{
vector
}
\description{
Asset contribution to portfolio standard deviation
}
\examples{
w <- c(.55, .25, .2)
s <- c(.24, .18, .15)
C <- matrix(c(1, 0.85, 0.3, 0.85, 1, -0.15, 0.3, -0.15, 1), nrow = 3)
asset_stdev_contrib(w, s, C)
asset_stdev_contrib(w, s, C, marginal = TRUE)
}
