% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_robust.R
\name{cor_robust}
\alias{cor_robust}
\title{Robustify a correlation matrix}
\usage{
cor_robust(C, e_thresh)
}
\arguments{
\item{C}{matrix}

\item{e_thresh}{vector}
}
\value{
matrix
}
\description{
Robustify a correlation matrix
}
