% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_robust.R
\name{cov_robust}
\alias{cov_robust}
\title{Robustify a covariance matrix}
\usage{
cov_robust(V, e_thresh)
}
\arguments{
\item{V}{matrix}

\item{e_thresh}{vector}
}
\value{
matrix
}
\description{
Robustify a covariance matrix
}
