% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_constituents.R
\name{effective_constituents}
\alias{effective_constituents}
\title{Title}
\usage{
effective_constituents(w, a = 2)
}
\arguments{
\item{w}{vector}

\item{a}{vector}
}
\value{
vector
}
\description{
Title
}
\examples{
effective_constituents(c(0.4, 0.4, 0.2))
}
