% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_stdev_contrib.R
\name{factor_stdev_contrib}
\alias{factor_stdev_contrib}
\title{Statistical factor contribution to portfolio standard deviation}
\usage{
factor_stdev_contrib(w, s, C)
}
\arguments{
\item{w}{matrix}

\item{s}{matrix}

\item{C}{matrix}
}
\value{
vector
}
\description{
Statistical factor contribution to portfolio standard deviation
}
