% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_diversified_portfolio.R
\name{most_diversified_portfolio}
\alias{most_diversified_portfolio}
\title{Most diversified portfolio of Choueifaty and Coignard (2008)}
\usage{
most_diversified_portfolio(s, C)
}
\arguments{
\item{s}{matrix of asset standard deviations}

\item{C}{matrix of asset correlations}
}
\value{
matrix of asset weights
}
\description{
Most diversified portfolio of Choueifaty and Coignard (2008)
}
\examples{
s <- matrix(c(.15, .3))
C <- matrix(c(1, .5, .5, 1), nrow = 2)
most_diversified_portfolio(s, C)
}
