% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_return.R
\name{portfolio_return}
\alias{portfolio_return}
\title{Portfolio return}
\usage{
portfolio_return(w, R)
}
\arguments{
\item{w}{vector}

\item{R}{matrix}
}
\value{
matrix
}
\description{
Portfolio return
}
\examples{
portfolio_return(c(0.5, 0.5), matrix(c(0.1, 0.2, 0.2, 0.1), ncol = 2))
}
