% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_stdev.R
\name{portfolio_stdev}
\alias{portfolio_stdev}
\title{Portfolio standard deviation}
\usage{
portfolio_stdev(w, s, C)
}
\arguments{
\item{w}{matrix}

\item{s}{matrix}

\item{C}{matrix}
}
\value{
vector
}
\description{
Portfolio standard deviation
}
\examples{
w <- matrix(c(.55, .25, .2))
s <- matrix(c(.24, .18, .15))
C <- matrix(c(1, 0.85, 0.3, 0.85, 1, -0.15, 0.3, -0.15, 1), nrow = 3)
portfolio_stdev(w, s, C)
}
