% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_return.R
\name{portfolio_return}
\alias{portfolio_return}
\title{Portfolio return}
\usage{
portfolio_return(w, r)
}
\arguments{
\item{w}{vector}

\item{r}{vector}
}
\value{
vector
}
\description{
Portfolio return
}
\examples{
w <- c(.55, .25, .2)
r <- c(0.08, 0.04, 0.03)
portfolio_return(w, r)
}
